# openapi.yaml
openapi: 3.0.0
info:
  title: 'Star Citizen API'
  contact:
    email: foxftw@star-citizen.wiki
  version: 3.0.0
servers:
  -
    url: 'https://api.star-citizen.wiki'
paths:
...
  /api/celestial-objects:
    get:
      tags:
        - Starmap
        - RSI-Website
      summary: 'Starmap Celestial Objects Overview'
      description: 'Returns paginated celestial objects with optional relationships.'
      operationId: cb1abb65b355a5cbf26f23c3b231a676
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
        -
          name: include
          in: query
          description: 'Include additional relationships (affiliation, starsystem).'
          explode: false
          allowReserved: true
          schema:
            type: string
      responses:
        '200':
          description: 'List of Celestial Objects'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/celestial_object'
  '/api/celestial-objects/{code}':
    get:
      tags:
        - Starmap
        - RSI-Website
      summary: 'Celestial Object Detail'
      description: 'Retrieve a celestial object by code or identifier, optionally including relations.'
      operationId: a56221ed3240e9c31bd996fb9cf85248
      parameters:
        -
          name: code
          in: path
          required: true
          schema:
            description: 'Celestial Object code or identifier'
            type: string
        -
          name: include
          in: query
          description: 'Include additional relationships (affiliation, starsystem).'
          explode: false
          allowReserved: true
          schema:
            type: string
      responses:
        '200':
          description: 'A Celestial Object'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/celestial_object'
        '404':
          description: 'No Celestial Object with specified code or name found.'
  /api/celestial-objects/search:
    post:
      tags:
        - Starmap
        - RSI-Website
        - Search
      summary: 'Celestial Object Search'
      description: 'Search celestial objects by code, cig_id, or name.'
      operationId: 34134c8a7a8bb87a6c6f905cbce9ed42
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
      requestBody:
        description: 'Partial celestial object code or name to search for'
        required: true
        content:
          application/json:
            schema:
              properties:
                query:
                  type: string
              type: object
            example: '{"query": "Pleiades"}'
      responses:
        '200':
          description: 'List of matching Celestial Objects'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/celestial_object'
...
components:
  schemas:
...
    celestial_object:
      title: 'Celestial Object'
      properties:
        id:
          type: integer
        code:
          type: string
        system_id:
          type: integer
        celestial_object_api_url:
          type: string
        name:
          type: string
        type:
          type: string
        age:
          type: integer
        habitable:
          type: boolean
        fairchanceact:
          type: boolean
        appearance:
          type: string
        designation:
          type: string
        distance:
          type: number
          format: float
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        axial_tilt:
          type: number
          format: float
        orbit_period:
          type: number
          format: float
        info_url:
          type: string
        description:
          type: object
        sensor:
          properties:
            population:
              type: number
              format: float
            economy:
              type: number
              format: float
            danger:
              type: number
              format: float
          type: object
        size:
          type: number
          format: float
        parent_id:
          type: integer
        time_modified:
          type: string
      type: object
...
  parameters:
...
    page:
      name: page
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_number:
      name: 'page[number]'
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: 'page[size]'
      in: query
      description: 'Number of results to return per page.'
      required: false
      schema:
        type: integer
        default: 30
        maximum: 200
        minimum: 1
...